refactor: moved TransactionDetailsInteractor
Moved interactor to /transaction package
