+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/interactor/transaction/TransactionDetailsInteractor.kt
@@ -14,7 +14,7 @@
  * governing permissions and limitations under the Licence.
  */
 
-package eu.europa.ec.issuancefeature.interactor.document
+package eu.europa.ec.issuancefeature.interactor.transaction
 
 import eu.europa.ec.commonfeature.ui.transaction_details.domain.TransactionClaimItem
 import eu.europa.ec.commonfeature.ui.transaction_details.domain.TransactionDetailsDomain
